+++ packages/common-testnets/src/index.ts
@@ -1,5 +1,5 @@
 export * from './TestnetClient.js'
 export * from './TestnetFactory.js'
-export * from './tenderly/index.js'
-export * from './anvil/index.js'
+export * from './nodes/tenderly/index.js'
+export * from './nodes/anvil/index.js'
 export * from './helpers/index.js'
