+++ packages/common-testnets/src/nodes/anvil/AnvilClient.ts
@@ -2,8 +2,8 @@ import { assert, Hash } from '@marsfoundation/common-universal'
 import { http, Address, createTestClient, numberToHex, publicActions, walletActions } from 'viem'
 import { dealActions } from 'viem-deal'
 import { mainnet } from 'viem/chains'
-import { TestnetClient } from '../TestnetClient.js'
-import { extendWithTestnetHelpers } from '../client-helpers.js'
+import { TestnetClient } from '../../TestnetClient.js'
+import { extendWithTestnetHelpers } from '../extendWithTestnetHelpers.js'
 
 export function getAnvilClient(rpc: string): TestnetClient {
   return createTestClient({
